add tests for normalize
